+++ cmd/conduit/root/pipelines/describe_test.go
@@ -17,13 +17,14 @@ package pipelines
 import (
 	"bytes"
 	"context"
+	"strings"
+	"testing"
+
 	"github.com/conduitio/conduit/cmd/conduit/api"
 	"github.com/conduitio/conduit/cmd/conduit/api/mock"
 	apiv1 "github.com/conduitio/conduit/proto/api/v1"
 	"github.com/conduitio/ecdysis"
 	"go.uber.org/mock/gomock"
-	"strings"
-	"testing"
 
 	"github.com/matryer/is"
 )
